+++ lib/ts/recipe/thirdpartyemailpassword/recipe.ts
@@ -45,6 +45,7 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
             RecipeImplementation({
                 recipeId: this.config.recipeId,
                 appInfo: this.config.appInfo,
+                clientType: this.config.clientType,
                 preAPIHook: this.config.preAPIHook,
                 postAPIHook: this.config.postAPIHook,
             })
@@ -61,6 +62,7 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
                 ? new EmailPasswordRecipe({
                       recipeId: this.config.recipeId,
                       appInfo: this.config.appInfo,
+                      clientType: this.config.clientType,
                       preAPIHook: config.preAPIHook,
                       postAPIHook: config.postAPIHook,
                       override: {
@@ -80,6 +82,7 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
                 ? new ThirdPartyRecipe({
                       recipeId: this.config.recipeId,
                       appInfo: this.config.appInfo,
+                      clientType: this.config.clientType,
                       preAPIHook: config.preAPIHook,
                       postAPIHook: config.postAPIHook,
                       override: {
@@ -104,12 +107,13 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
     }
 
     static init(config?: UserInput): CreateRecipeFunction<PreAndPostAPIHookAction> {
-        return (appInfo: NormalisedAppInfo) => {
+        return (appInfo: NormalisedAppInfo, clientType: string | undefined) => {
             Recipe.instance = new Recipe(
                 {
                     ...config,
                     recipeId: Recipe.RECIPE_ID,
                     appInfo,
+                    clientType,
                 },
                 {
                     emailPassword: undefined,