+++ lib/ts/recipe/thirdpartypasswordless/recipe.ts
@@ -59,6 +59,7 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
                 ? new ThirdPartyRecipe({
                       recipeId: this.config.recipeId,
                       appInfo: this.config.appInfo,
+                      clientType: this.config.clientType,
                       preAPIHook: config.preAPIHook,
                       postAPIHook: config.postAPIHook,
                       override: {
@@ -74,6 +75,7 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
                 ? new PasswordlessRecipe({
                       recipeId: this.config.recipeId,
                       appInfo: this.config.appInfo,
+                      clientType: this.config.clientType,
                       preAPIHook: config.preAPIHook,
                       postAPIHook: config.postAPIHook,
                       override: {
@@ -98,12 +100,13 @@ export default class Recipe extends AuthRecipe<PreAndPostAPIHookAction, Normalis
     }
 
     static init(config?: UserInput): CreateRecipeFunction<PreAndPostAPIHookAction> {
-        return (appInfo: NormalisedAppInfo) => {
+        return (appInfo: NormalisedAppInfo, clientType: string | undefined) => {
             Recipe.instance = new Recipe(
                 {
                     ...config,
                     recipeId: Recipe.RECIPE_ID,
                     appInfo,
+                    clientType,
                 },
                 {
                     thirdParty: undefined,