+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.7.2] - 2023-08-07
+
+### Fixes
+
+-   Fixed `clientType` usage in `thirdpartyemailpassword` and `thirdpartypasswordless`
+
 ## [0.7.1] - 2023-07-31
 
 ### Changes